refactor: fake menu policy 수정
